requirements.txt
# Requirements for Llama-2 QLoRA fine-tuning

# Core ML libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
pytorch-lightning>=2.0.0

# QLoRA and PEFT
peft>=0.6.0
bitsandbytes>=0.41.0

# Additional dependencies
sentencepiece>=0.1.99
protobuf>=3.20.3
hydra-core>=1.3.0
omegaconf>=2.3.0

# Monitoring and utilities
codecarbon>=2.3.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0

# Optional but recommended
huggingface-hub>=0.17.0
tokenizers>=0.14.0

# For Colab users
ipywidgets>=8.0.0
